Move history related stuff to separate class
